MensajeMontoCuenta

+++ BancoMBPersistencias/src/main/java/org/itson/bda/proyectobda_247164_246943/daos/CuentasDAO.java
@@ -79,7 +79,7 @@ public class CuentasDAO implements ICuentasDAO {
         try (
                 Connection conexion = this.conexionBD.obtenerConexion(); PreparedStatement comando = conexion.prepareStatement(sentenciaSQL, Statement.RETURN_GENERATED_KEYS);) {
             comando.setDate(1, cuentaNueva.getFechaApertura());
-            comando.setFloat(2, cuentaNueva.getSaldo());
+            comando.setDouble(2, cuentaNueva.getSaldo());
             int numeroRegistrosInsertados = comando.executeUpdate();
             logger.log(Level.INFO, "Se agregaron {0} cuentas", numeroRegistrosInsertados);
             ResultSet numeroCuentasGeneradas = comando.getGeneratedKeys();
